requirements.txt
# Core dependencies for Multi-Agent Financial System
langchain>=0.1.0
langchain-community>=0.1.0
langchain-groq>=0.0.1
langgraph>=0.0.26

# Vector Database
pinecone-client>=3.0.0

# Financial Data
alpha-vantage>=2.3.1

# Web Search
duckduckgo-search>=4.0.0

# Embeddings
sentence-transformers>=2.2.2

# Data Processing
pandas>=2.0.0
tabulate>=0.9.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
